feat: Update SummaryWorkoutBottomSheet onPressed logic
The `onPressed` logic in the `SummaryWorkoutBottomSheet` widget has been updated to handle authentication and saving records. If the user is not logged in, they will be prompted to login with Google. Once logged in, the `totalJumpingJack` count will be saved as a record and the `RecordsBottomSheet` will be displayed. This change improves the user experience and allows for tracking and displaying workout progress.